Fix under toolbar in SwiftUI layout. Fix swift previews by using ios 15 for every framework which shouldn't be necessary, but SwiftUI is trying to build whole app.
