--- catowser/catowser/Browser/Previews/MainBrowserViewPreview.swift
@@ -1,18 +1,0 @@
-//
-//  MainBrowserViewPreview.swift
-//  catowser
-//
-//  Created by Andrey Ermoshin on 10.12.2022.
-//  Copyright © 2022 andreiermoshin. All rights reserved.
-//
-
-import SwiftUI
-
-#if DEBUG
-// swiftlint:disable type_name
-struct MainBrowserView_Previews: PreviewProvider {
-    static var previews: some View {
-        MainBrowserView()
-    }
-}
-#endif
