+++ catowser/catowser/Theme/ThemeProvider.swift
@@ -8,11 +8,6 @@
 
 import UIKit
 
-extension UIColor {
-    static let phoneToolbarColor: UIColor = #colorLiteral(red: 0.9647058824, green: 0.9647058824, blue: 0.9647058824, alpha: 1)
-    static let linkTagsBackgroundColor = #colorLiteral(red: 0.8039215803, green: 0.8039215803, blue: 0.8039215803, alpha: 1)
-}
-
 final class ThemeProvider {
     static let shared = ThemeProvider()
