+++ catowser/catowser/Theme/UIColor+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  UIColor+Extension.swift
+//  catowser
+//
+//  Created by Andrei Ermoshin on 12/13/22.
+//  Copyright © 2022 andreiermoshin. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    static let phoneToolbarColor: UIColor = #colorLiteral(red: 0.9647058824, green: 0.9647058824, blue: 0.9647058824, alpha: 1)
+    static let linkTagsBackgroundColor = #colorLiteral(red: 0.8039215803, green: 0.8039215803, blue: 0.8039215803, alpha: 1)
+}
